renormalization

+++ lbry/wallet/server/db/elasticsearch/constants.py
@@ -30,7 +30,7 @@ INDEX_DEFAULT_SETTINGS = {
             "height": {"type": "integer"},
             "claim_type": {"type": "byte"},
             "censor_type": {"type": "byte"},
-            "trending_score": {"type": "float"},
+            "trending_score": {"type": "double"},
             "release_time": {"type": "long"}
         }
     }
